Rename files in dsl-mastery
