+++ .changeset/tender-peaches-pump.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-extension-dsl-mastery': major
+---
+
+Rename files prefix from `DSLMastery` to `DSL_Mastery`.
