+++ packages/legend-extension-dsl-mastery/src/components/studio/DSL_Mastery_CodeSnippets.ts
