+++ packages/legend-extension-dsl-mastery/src/graph/metamodel/pure/model/packageableElements/mastery/DSL_Mastery_IdentityResolution.ts
@@ -16,7 +16,7 @@
 
 import { type Hashable, hashArray } from '@finos/legend-shared';
 import type { RawLambda } from '@finos/legend-graph';
-import { MASTERY_HASH_STRUCTURE } from '../../../../../DSLMastery_HashUtils.js';
+import { MASTERY_HASH_STRUCTURE } from '../../../../../DSL_Mastery_HashUtils.js';
 
 export class IdentityResolution implements Hashable {
   modelClass!: string;
