+++ packages/legend-extension-dsl-mastery/src/graphManager/__tests__/DSL_Mastery_Roundtrip.test.ts
@@ -15,8 +15,8 @@
  */
 
 import { test } from '@jest/globals';
-import { DSLMastery_GraphManagerPreset } from '../../DSLMastery_Extension.js';
-import { TEST_DATA__roundtrip } from './TEST_DATA__DSLMastery_Roundtrip.js';
+import { DSL_Mastery_GraphManagerPreset } from '../../DSL_Mastery_Extension.js';
+import { TEST_DATA__roundtrip } from './TEST_DATA__DSL_Mastery_Roundtrip.js';
 import {
   DSLExternalFormat_GraphPreset,
   TEST__checkBuildingElementsRoundtrip,
@@ -28,7 +28,7 @@ import type { Entity } from '@finos/legend-storage';
 const pluginManager = new TEST__GraphManagerPluginManager();
 pluginManager
   .usePresets([
-    new DSLMastery_GraphManagerPreset(),
+    new DSL_Mastery_GraphManagerPreset(),
     new DSLExternalFormat_GraphPreset(),
   ])
   .install();
