+++ packages/legend-extension-dsl-mastery/src/graphManager/protocol/pure/v1/transformation/pureGraph/from/V1_DSL_Mastery_TransformerHelper.ts
@@ -14,24 +14,24 @@
  * limitations under the License.
  */
 
-import type { MasterRecordDefinition } from '../../../../../../../graph/metamodel/pure/model/packageableElements/mastery/DSLMastery_MasterRecordDefinition.js';
+import type { MasterRecordDefinition } from '../../../../../../../graph/metamodel/pure/model/packageableElements/mastery/DSL_Mastery_MasterRecordDefinition.js';
 import type {
   IdentityResolution,
   ResolutionQuery,
-} from '../../../../../../../graph/metamodel/pure/model/packageableElements/mastery/DSLMastery_IdentityResolution.js';
+} from '../../../../../../../graph/metamodel/pure/model/packageableElements/mastery/DSL_Mastery_IdentityResolution.js';
 import type {
   RecordSource,
   RecordSourcePartition,
-} from '../../../../../../../graph/metamodel/pure/model/packageableElements/mastery/DSLMastery_RecordSource.js';
-import { V1_MasterRecordDefinition } from '../../../model/packageableElements/mastery/V1_DSLMastery_MasterRecordDefinition.js';
+} from '../../../../../../../graph/metamodel/pure/model/packageableElements/mastery/DSL_Mastery_RecordSource.js';
+import { V1_MasterRecordDefinition } from '../../../model/packageableElements/mastery/V1_DSL_Mastery_MasterRecordDefinition.js';
 import {
   V1_IdentityResolution,
   V1_ResolutionQuery,
-} from '../../../model/packageableElements/mastery/V1_DSLMastery_IdentityResolution.js';
+} from '../../../model/packageableElements/mastery/V1_DSL_Mastery_IdentityResolution.js';
 import {
   V1_RecordSource,
   V1_RecordSourcePartition,
-} from '../../../model/packageableElements/mastery/V1_DSLMastery_RecordSource.js';
+} from '../../../model/packageableElements/mastery/V1_DSL_Mastery_RecordSource.js';
 import {
   type V1_GraphTransformerContext,
   V1_initPackageableElement,
