+++ packages/legend-manual-tests/src/__tests__/roundtrip-grammar/RoundtripGrammar.test.ts
@@ -51,7 +51,7 @@ import { DSL_Text_GraphManagerPreset } from '@finos/legend-extension-dsl-text';
 import { DSL_Diagram_GraphManagerPreset as DSL_Diagram_GraphManagerPreset } from '@finos/legend-extension-dsl-diagram';
 import { DSL_DataSpace_GraphManagerPreset } from '@finos/legend-extension-dsl-data-space';
 import { DSLPersistence_GraphManagerPreset } from '@finos/legend-extension-dsl-persistence';
-import { DSLMastery_GraphManagerPreset } from '@finos/legend-extension-dsl-mastery';
+import { DSL_Mastery_GraphManagerPreset } from '@finos/legend-extension-dsl-mastery';
 import { DSLPersistenceCloud_GraphManagerPreset } from '@finos/legend-extension-dsl-persistence-cloud';
 import { ESService_GraphManagerPreset } from '@finos/legend-extension-external-store-service';
 
@@ -133,7 +133,7 @@ const checkGrammarRoundtrip = async (
       new DSL_DataSpace_GraphManagerPreset(),
       new DSLExternalFormat_GraphPreset(),
       new DSLPersistence_GraphManagerPreset(),
-      new DSLMastery_GraphManagerPreset(),
+      new DSL_Mastery_GraphManagerPreset(),
       new DSLPersistenceCloud_GraphManagerPreset(),
       new ESService_GraphManagerPreset(),
     ])
